Add support for custom IN and OUT labels
Added command line argument support to change the IN and OUT labels so people can use a custom text.

The command line arguments are -i/--inlabel and -o/--outlabel respectively.

This is useful if someone would like to display up and down unicode arrows for example.

If no values are provided the script behaves exactly as before, note that the hardcoded spaces between the IN/OUT label and the speed have been removed. To keep the script working as before when no arguments are passed the default values are "IN " and "OUT " with a space, but the user should type the space character manually if setting a custom label. This is done to allow more flexibility to the script, should anyone need it.